Bitcoin Miner System
Bitcoins (see http://en.wikipedia.org/wiki/Bitcoin) are the most
popular crypto-currency in common use. At their hart, bitcoins use the
hardness of cryp-tographic hashing (for a reference see
http://en.wikipedia.org/wiki/Cryptographic hash function) to ensure a
limited “supply” of coins. In particular, the key component in a
bit-coin is an input that, when “hashed” produces an output smaller
than a target value. In practice, the comparison values have leading
0’s, thus the bitcoin is
required to have a given number of leading 0’s (to ensure 3 leading
0’s, you look for hashes smaller than 0x001000... or smaller or equal
to 0x000ff....
The hash i used is SHA-256. You can check your version against this
online hasher: http:/www.xorbin.com/tools/sha256-hash-calculator.For
the coins you find, check your answer with this calculator to ensure
correctness.
The goal of this  project is to use Scala and the actor model to build
a good solution to this problem that runs well on multi-core machines.